+++ src/Krystal/Db/Sql/AbstractMapper.php
@@ -307,6 +307,49 @@ abstract class AbstractMapper
     }
 
     /**
+     * Returns a sum of columns
+     * 
+     * @param array $columns
+     * @param array $constraints
+     * @param integer $precision Float precision
+     * @return array
+     */
+    final public function getColumsSum(array $columns, array $constraints = array(), $precision = 2)
+    {
+        $db = $this->db->select();
+
+        foreach ($columns as $column) {
+            $db->sum($column, $column);
+        }
+
+        $db->from(static::getTableName());
+
+        if (!empty($constraints)) {
+            // Iteration counter
+            $iteration = 0;
+
+            foreach ($constraints as $key => $value) {
+                if ($iteration === 0) {
+                    $db->whereEquals($key, $value);
+                } else {
+                    $db->andWhereEquals($key, $value);
+                }
+
+                $iteration++;
+            }
+        }
+
+        $data = $db->queryAll();
+
+        if (isset($data[0])) {
+            return Math::roundCollection($data[0], $precision);
+        } else {
+            // No results
+            return array();
+        }
+    }
+
+    /**
      * Persists a row
      * 
      * @param array $data